requirements.txt
# Django
Django==2.1.11
pytz==2019.2

# RESTful framework
djangorestframework==3.10.2
